Add github package for downloading releases
